requirements.txt
fastapi~=0.112.2
pdfminer-six~=20191110
openai~=1.51.2
python-dotenv~=1.0.1
pydantic~=2.8.2